[BH-741] Split ApplicationManager
Separate Pure and Bell ApplicationManager functionalities
to separate files.
