+++ module-apps/application-desktop/windows/ChargingBatteryWindow.cpp
@@ -8,7 +8,6 @@
 #include <gui/widgets/Image.hpp>
 #include <log.hpp>
 #include <service-appmgr/Controller.hpp>
-#include <service-appmgr/model/ApplicationManager.hpp>
 
 namespace gui
 {
