+++ module-apps/apps-common/windows/OptionWindow.cpp
@@ -5,7 +5,6 @@
 #include "Label.hpp"
 #include <i18n/i18n.hpp>
 #include <log.hpp>
-#include <service-appmgr/model/ApplicationManager.hpp>
 #include <Style.hpp>
 #include <memory>
 #include <utility>
