+++ module-services/service-appmgr/include/service-appmgr/Constants.hpp
@@ -1,0 +1,9 @@
+// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+// For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
+
+#pragma once
+
+namespace service::name
+{
+    inline constexpr auto appmgr = "ApplicationManager";
+};
