+++ module-services/service-appmgr/include/service-appmgr/model/ActionsRegistry.hpp
@@ -42,7 +42,8 @@ namespace app::manager
     {
         Accepted,
         Skipped,
-        Dropped
+        Dropped,
+        NotHandled
     };
 
     class ActionsRegistry
