+++ module-services/service-appmgr/model/ApplicationManagerCommon.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
-#include "ApplicationManager.hpp"
+#include "ApplicationManagerCommon.hpp"
 #include "ApplicationStatus.hpp"
 #include "AutoLockRequests.hpp"
 #include "Controller.hpp"
@@ -31,6 +31,7 @@
 #include <module-services/service-db/agents/settings/SystemSettings.hpp>
 #include <module-sys/Timers/TimerFactory.hpp>
 #include <module-utils/Utils.hpp>
+#include <service-appmgr/Constants.hpp>
 #include <service-appmgr/StartupType.hpp>
 #include <service-cellular-api>
 #include <service-cellular/CellularMessage.hpp>
@@ -49,7 +50,6 @@ namespace app::manager
     namespace
     {
         constexpr auto ApplicationManagerStackDepth = 3072;
-        constexpr auto autoLockTimerName            = "AutoLockTimer";
     } // namespace
 
     ApplicationManagerBase::ApplicationManagerBase(std::vector<std::unique_ptr<app::ApplicationLauncher>> &&launchers)
@@ -116,49 +116,24 @@ namespace app::manager
         return runningApps;
     }
 
-    ApplicationManager::ApplicationManager(const ApplicationName &serviceName,
-                                           std::vector<std::unique_ptr<app::ApplicationLauncher>> &&launchers,
-                                           const ApplicationName &_rootApplicationName)
+    ApplicationManagerCommon::ApplicationManagerCommon(
+        const ApplicationName &serviceName,
+        std::vector<std::unique_ptr<app::ApplicationLauncher>> &&launchers,
+        const ApplicationName &_rootApplicationName)
         : Service{serviceName, {}, ApplicationManagerStackDepth},
           ApplicationManagerBase(std::move(launchers)), rootApplicationName{_rootApplicationName},
           actionsRegistry{[this](ActionEntry &action) { return handleAction(action); }},
-          settings(std::make_shared<settings::Settings>()),
-          phoneModeObserver(std::make_shared<sys::phone_modes::Observer>()),
-          phoneLockHandler(locks::PhoneLockHandler(this, settings)),
-          simLockHandler(this), notificationsConfig{phoneModeObserver, settings, phoneLockHandler},
-          notificationsHandler{this, notificationsConfig}, notificationProvider{this, notificationsConfig}
+          settings(std::make_shared<settings::Settings>())
     {
-        autoLockTimer = sys::TimerFactory::createSingleShotTimer(
-            this, autoLockTimerName, sys::timer::InfiniteTimeout, [this](sys::Timer &) { onPhoneLocked(); });
-        bus.channels.push_back(sys::BusChannel::PhoneModeChanges);
         bus.channels.push_back(sys::BusChannel::ServiceAudioNotifications);
         bus.channels.push_back(sys::BusChannel::ServiceDBNotifications);
-        bus.channels.push_back(sys::BusChannel::PhoneLockChanges);
-        bus.channels.push_back(sys::BusChannel::ServiceCellularNotifications);
-        registerMessageHandlers();
     }
 
-    sys::ReturnCodes ApplicationManager::InitHandler()
+    sys::ReturnCodes ApplicationManagerCommon::InitHandler()
     {
         settings->init(service::ServiceProxy(shared_from_this()));
         utils::setDisplayLanguage(
             settings->getValue(settings::SystemProperties::displayLanguage, settings::SettingsScope::Global));
-
-        phoneLockHandler.enablePhoneLock((utils::getNumericValue<bool>(
-            settings->getValue(settings::SystemProperties::lockScreenPasscodeIsOn, settings::SettingsScope::Global))));
-
-        phoneLockHandler.setPhoneLockHash(
-            settings->getValue(settings::SystemProperties::lockPassHash, settings::SettingsScope::Global));
-
-        settings->registerValueChange(
-            settings::SystemProperties::lockScreenPasscodeIsOn,
-            [this](const std::string &value) { phoneLockHandler.enablePhoneLock(utils::getNumericValue<bool>(value)); },
-            settings::SettingsScope::Global);
-
-        settings->registerValueChange(
-            settings::SystemProperties::lockPassHash,
-            [this](const std::string &value) { phoneLockHandler.setPhoneLockHash(value); },
-            settings::SettingsScope::Global);
 
         settings->registerValueChange(
             settings::SystemProperties::displayLanguage,
@@ -168,26 +143,13 @@ namespace app::manager
             settings::SystemProperties::inputLanguage,
             [this](std::string value) { inputLanguageChanged(std::move(value)); },
             settings::SettingsScope::Global);
-        settings->registerValueChange(
-            settings::SystemProperties::autoLockTimeInSec,
-            [this](std::string value) { lockTimeChanged(std::move(value)); },
-            settings::SettingsScope::Global);
-        settings->registerValueChange(
-            ::settings::KeypadLight::state,
-            [this](const std::string &value) {
-                const auto keypadLightState =
-                    static_cast<bsp::keypad_backlight::State>(utils::getNumericValue<int>(value));
-                processKeypadBacklightState(keypadLightState);
-            },
-            ::settings::SettingsScope::Global);
 
-        startBackgroundApplications();
         bus.sendUnicast(std::make_unique<CheckIfStartAllowedMessage>(), service::name::system_manager);
 
         return sys::ReturnCodes::Success;
     }
 
-    void ApplicationManager::handleStart(StartAllowedMessage *msg)
+    void ApplicationManagerCommon::handleStart(StartAllowedMessage *msg)
     {
         switch (msg->getStartupType()) {
         case StartupType::Regular:
@@ -196,41 +158,33 @@ namespace app::manager
             }
             break;
         case StartupType::LowBattery:
-            handleSwitchApplication(
-                std::make_unique<SwitchRequest>(ServiceName, app::name_desktop, window::name::dead_battery, nullptr)
-                    .get());
+            handleSwitchApplication(std::make_unique<SwitchRequest>(
+                                        service::name::appmgr, app::name_desktop, window::name::dead_battery, nullptr)
+                                        .get());
             break;
         case StartupType::LowBatteryCharging:
             handleSwitchApplication(
-                std::make_unique<SwitchRequest>(ServiceName, app::name_desktop, window::name::charging_battery, nullptr)
+                std::make_unique<SwitchRequest>(
+                    service::name::appmgr, app::name_desktop, window::name::charging_battery, nullptr)
                     .get());
             break;
         }
     }
 
-    void ApplicationManager::suspendSystemServices()
+    void ApplicationManagerCommon::suspendSystemServices()
     {
         sys::SystemManager::SuspendService(service::name::gui, this);
         sys::SystemManager::SuspendService(service::name::eink, this);
     }
 
-    void ApplicationManager::startBackgroundApplications()
-    {
-        for (const auto &name : std::vector<ApplicationName>{app::special_input}) {
-            if (auto app = getApplication(name); app != nullptr) {
-                app->runInBackground(phoneModeObserver->getCurrentPhoneMode(), this);
-            }
-        }
-    }
-
-    sys::ReturnCodes ApplicationManager::DeinitHandler()
+    sys::ReturnCodes ApplicationManagerCommon::DeinitHandler()
     {
         settings->deinit();
         closeApplications();
         return sys::ReturnCodes::Success;
     }
 
-    auto ApplicationManager::ProcessCloseReason(sys::CloseReason closeReason) -> void
+    auto ApplicationManagerCommon::ProcessCloseReason(sys::CloseReason closeReason) -> void
     {
         ActionRequest act = ActionRequest{this->GetName(), app::manager::actions::DisplayLogoAtExit, nullptr};
         switch (closeReason) {
@@ -248,26 +202,15 @@ namespace app::manager
         sendCloseReadyMessage(this);
     }
 
-    auto ApplicationManager::DataReceivedHandler([[maybe_unused]] sys::DataMessage *msgl,
-                                                 [[maybe_unused]] sys::ResponseMessage *resp) -> sys::MessagePointer
+    auto ApplicationManagerCommon::DataReceivedHandler([[maybe_unused]] sys::DataMessage *msgl,
+                                                       [[maybe_unused]] sys::ResponseMessage *resp)
+        -> sys::MessagePointer
     {
         return std::make_shared<sys::ResponseMessage>();
     }
 
-    void ApplicationManager::registerMessageHandlers()
+    void ApplicationManagerCommon::registerMessageHandlers()
     {
-        phoneModeObserver->connect(this);
-        phoneModeObserver->subscribe([this](sys::phone_modes::PhoneMode phoneMode) {
-            handlePhoneModeChanged(phoneMode);
-            actionsRegistry.enqueue(
-                ActionEntry{actions::ShowPopup, std::make_unique<gui::PhoneModePopupRequestParams>(phoneMode)});
-        });
-
-        phoneModeObserver->subscribe(
-            [this](sys::phone_modes::Tethering tethering) { handleTetheringChanged(tethering); });
-
-        notificationsHandler.registerMessageHandlers();
-
         connect(typeid(StartAllowedMessage), [this](sys::Message *request) {
             auto msg = static_cast<StartAllowedMessage *>(request);
             handleStart(msg);
@@ -280,12 +223,6 @@ namespace app::manager
         });
         connect(typeid(PowerSaveModeInitRequest), [this]([[maybe_unused]] sys::Message *msg) {
             handlePowerSavingModeInit();
-            return std::make_shared<sys::ResponseMessage>();
-        });
-        connect(typeid(PreventBlockingRequest), [this]([[maybe_unused]] sys::Message *msg) {
-            if (!phoneLockHandler.isPhoneLocked()) {
-                autoLockTimer.start();
-            }
             return std::make_shared<sys::ResponseMessage>();
         });
         connect(typeid(SwitchRequest), [this](sys::Message *request) {
@@ -350,176 +287,6 @@ namespace app::manager
         connect(typeid(GetCurrentDisplayLanguageRequest), [&](sys::Message *request) {
             return std::make_shared<GetCurrentDisplayLanguageResponse>(utils::getDisplayLanguage());
         });
-        connect(typeid(GetAllNotificationsRequest), [&](sys::Message *request) {
-            notificationProvider.requestNotSeenNotifications();
-            notificationProvider.send();
-            return sys::msgHandled();
-        });
-        connect(typeid(db::NotificationMessage), [&](sys::Message *msg) {
-            auto msgl = static_cast<db::NotificationMessage *>(msg);
-            notificationProvider.handle(msgl);
-            return sys::msgHandled();
-        });
-        connect(typeid(db::QueryResponse), [&](sys::Message *msg) {
-            auto response = static_cast<db::QueryResponse *>(msg);
-            handleDBResponse(response);
-            return sys::msgHandled();
-        });
-
-        // PhoneLock connects
-        connect(typeid(locks::LockPhone),
-                [&](sys::Message *request) -> sys::MessagePointer { return phoneLockHandler.handleLockRequest(); });
-        connect(typeid(locks::UnlockPhone),
-                [&](sys::Message *request) -> sys::MessagePointer { return phoneLockHandler.handleUnlockRequest(); });
-        connect(typeid(locks::CancelUnlockPhone), [&](sys::Message *request) -> sys::MessagePointer {
-            return phoneLockHandler.handleUnlockCancelRequest();
-        });
-        connect(typeid(locks::UnLockPhoneInput), [&](sys::Message *request) -> sys::MessagePointer {
-            auto data = static_cast<locks::UnLockPhoneInput *>(request);
-            return phoneLockHandler.handlePhoneLockInput(data->getInputData());
-        });
-        connect(typeid(locks::EnablePhoneLock),
-                [&](sys::Message *request) -> sys::MessagePointer { return phoneLockHandler.handleEnablePhoneLock(); });
-        connect(typeid(locks::DisablePhoneLock), [&](sys::Message *request) -> sys::MessagePointer {
-            return phoneLockHandler.handleDisablePhoneLock();
-        });
-        connect(typeid(locks::UnlockedPhone), [&](sys::Message *request) -> sys::MessagePointer {
-            autoLockTimer.start();
-            return simLockHandler.releaseSimUnlockBlockOnLockedPhone();
-        });
-        connect(typeid(locks::ChangePhoneLock),
-                [&](sys::Message *request) -> sys::MessagePointer { return phoneLockHandler.handleChangePhoneLock(); });
-        connect(typeid(locks::SetPhoneLock),
-                [&](sys::Message *request) -> sys::MessagePointer { return phoneLockHandler.handleSetPhoneLock(); });
-        connect(typeid(locks::SkipSetPhoneLock), [&](sys::Message *request) -> sys::MessagePointer {
-            return phoneLockHandler.handleSkipSetPhoneLock();
-        });
-        connect(typeid(GetAutoLockTimeoutRequest), [&](sys::Message *request) -> sys::MessagePointer {
-            auto req = static_cast<GetAutoLockTimeoutRequest *>(request);
-            return handleAutoLockGetRequest(req);
-        });
-        connect(typeid(SetAutoLockTimeoutRequest), [&](sys::Message *request) -> sys::MessagePointer {
-            auto req = static_cast<SetAutoLockTimeoutRequest *>(request);
-            return handleAutoLockSetRequest(req);
-        });
-        connect(typeid(locks::ExternalUnLockPhone), [&](sys::Message *request) -> sys::MessagePointer {
-            auto data = static_cast<locks::ExternalUnLockPhone *>(request);
-            return phoneLockHandler.handleExternalUnlockRequest(data->getInputData());
-        });
-        connect(typeid(locks::ExternalPhoneLockAvailabilityChange), [&](sys::Message *request) -> sys::MessagePointer {
-            auto data = static_cast<locks::ExternalPhoneLockAvailabilityChange *>(request);
-            return phoneLockHandler.handleExternalAvailabilityChange(data->getAvailability());
-        });
-
-        // SimLock connects
-        connect(typeid(cellular::msg::notification::SimNeedPin), [&](sys::Message *request) -> sys::MessagePointer {
-            auto data = static_cast<cellular::msg::notification::SimNeedPin *>(request);
-            if (phoneLockHandler.isPhoneLocked()) {
-                simLockHandler.setSimUnlockBlockOnLockedPhone();
-            }
-            return simLockHandler.handleSimPinRequest(data->attempts);
-        });
-        connect(typeid(cellular::msg::request::sim::PinUnlock::Response),
-                [&](sys::Message *request) -> sys::MessagePointer {
-                    auto data = static_cast<cellular::msg::request::sim::PinUnlock::Response *>(request);
-                    if (data->retCode) {
-                        return simLockHandler.handleSimUnlockedMessage();
-                    }
-                    return sys::msgNotHandled();
-                });
-        connect(typeid(locks::UnLockSimInput), [&](sys::Message *request) -> sys::MessagePointer {
-            auto msg = static_cast<locks::UnLockSimInput *>(request);
-            return simLockHandler.verifySimLockInput(msg->getInputData());
-        });
-        connect(typeid(cellular::msg::notification::SimNeedPuk), [&](sys::Message *request) -> sys::MessagePointer {
-            auto data = static_cast<cellular::msg::notification::SimNeedPuk *>(request);
-            if (phoneLockHandler.isPhoneLocked()) {
-                simLockHandler.setSimUnlockBlockOnLockedPhone();
-            }
-            return simLockHandler.handleSimPukRequest(data->attempts);
-        });
-        connect(typeid(cellular::msg::request::sim::UnblockWithPuk::Response),
-                [&](sys::Message *request) -> sys::MessagePointer {
-                    auto data = static_cast<cellular::msg::request::sim::UnblockWithPuk::Response *>(request);
-                    if (data->retCode) {
-                        return simLockHandler.handleSimUnlockedMessage();
-                    }
-                    return sys::msgNotHandled();
-                });
-        connect(typeid(locks::ChangeSimPin), [&](sys::Message *request) -> sys::MessagePointer {
-            return simLockHandler.handleSimPinChangeRequest();
-        });
-        connect(typeid(cellular::msg::request::sim::ChangePin::Response),
-                [&](sys::Message *request) -> sys::MessagePointer {
-                    auto data = static_cast<cellular::msg::request::sim::ChangePin::Response *>(request);
-                    if (data->retCode) {
-                        return simLockHandler.handleSimPinChangedMessage();
-                    }
-                    else {
-                        return simLockHandler.handleSimPinChangeFailedRequest();
-                    }
-                });
-        connect(typeid(locks::EnableSimPin),
-                [&](sys::Message *request) -> sys::MessagePointer { return simLockHandler.handleSimEnableRequest(); });
-        connect(typeid(locks::DisableSimPin),
-                [&](sys::Message *request) -> sys::MessagePointer { return simLockHandler.handleSimDisableRequest(); });
-        connect(typeid(cellular::msg::request::sim::SetPinLock::Response),
-                [&](sys::Message *request) -> sys::MessagePointer {
-                    auto data = static_cast<cellular::msg::request::sim::SetPinLock::Response *>(request);
-                    if (data->retCode) {
-                        return simLockHandler.handleSimAvailabilityMessage();
-                    }
-                    else {
-                        if (data->lock == cellular::api::SimLockState::Enabled) {
-                            return simLockHandler.handleSimEnableRequest();
-                        }
-                        else {
-                            return simLockHandler.handleSimDisableRequest();
-                        }
-                    }
-                });
-        connect(typeid(cellular::msg::notification::SimBlocked), [&](sys::Message *request) -> sys::MessagePointer {
-            if (phoneLockHandler.isPhoneLocked()) {
-                simLockHandler.setSimUnlockBlockOnLockedPhone();
-            }
-            return simLockHandler.handleSimBlockedRequest();
-        });
-        connect(typeid(cellular::msg::notification::UnhandledCME), [&](sys::Message *request) -> sys::MessagePointer {
-            auto data = static_cast<cellular::msg::notification::UnhandledCME *>(request);
-            if (phoneLockHandler.isPhoneLocked()) {
-                simLockHandler.setSimUnlockBlockOnLockedPhone();
-            }
-            return simLockHandler.handleCMEErrorRequest(data->code);
-        });
-        connect(typeid(locks::SetSim), [&](sys::Message *request) -> sys::MessagePointer {
-            auto data = static_cast<locks::SetSim *>(request);
-            simLockHandler.setSim(data->getSimSlot());
-            return sys::msgHandled();
-        });
-        connect(typeid(cellular::msg::request::sim::SetActiveSim::Response),
-                [&](sys::Message *request) -> sys::MessagePointer {
-                    auto data = static_cast<cellular::msg::request::sim::SetActiveSim::Response *>(request);
-                    if (data->retCode) {
-                        settings->setValue(::settings::SystemProperties::activeSim,
-                                           utils::enumToString(Store::GSM::get()->selected),
-                                           ::settings::SettingsScope::Global);
-                        return sys::msgHandled();
-                    }
-                    return sys::msgNotHandled();
-                });
-        connect(typeid(cellular::msg::notification::SimReady),
-                [&](sys::Message *request) -> sys::MessagePointer { return simLockHandler.handleSimReadyMessage(); });
-        connect(typeid(cellular::msg::notification::ModemStateChanged),
-                [&](sys::Message *request) -> sys::MessagePointer {
-                    auto data = static_cast<cellular::msg::notification::ModemStateChanged *>(request);
-                    if (data->state == cellular::api::ModemState::Ready) {
-                        simLockHandler.setSimReady();
-                        simLockHandler.getSettingsSimSelect(
-                            settings->getValue(settings::SystemProperties::activeSim, settings::SettingsScope::Global));
-                        return sys::msgHandled();
-                    }
-                    return sys::msgNotHandled();
-                });
 
         connect(typeid(onBoarding::FinalizeOnBoarding),
                 [&](sys::Message *request) -> sys::MessagePointer { return handleOnBoardingFinalize(); });
@@ -530,26 +297,11 @@ namespace app::manager
         connect(typeid(app::manager::DOMRequest), [&](sys::Message *request) { return handleDOMRequest(request); });
 
         auto convertibleToActionHandler = [this](sys::Message *request) { return handleMessageAsAction(request); };
-        connect(typeid(CellularMMIResultMessage), convertibleToActionHandler);
-        connect(typeid(CellularMMIResponseMessage), convertibleToActionHandler);
-        connect(typeid(CellularMMIPushMessage), convertibleToActionHandler);
-        connect(typeid(CellularNoSimNotification), convertibleToActionHandler);
-        connect(typeid(CellularNotAnEmergencyNotification), convertibleToActionHandler);
         connect(typeid(sys::CriticalBatteryLevelNotification), convertibleToActionHandler);
-        connect(typeid(CellularSmsNoSimRequestMessage), convertibleToActionHandler);
-        connect(typeid(CellularSMSRejectedByOfflineNotification), convertibleToActionHandler);
-        connect(typeid(CellularCallRejectedByOfflineNotification), convertibleToActionHandler);
-        connect(typeid(sys::TetheringQuestionRequest), convertibleToActionHandler);
-        connect(typeid(sys::TetheringQuestionAbort), convertibleToActionHandler);
-        connect(typeid(sys::TetheringPhoneModeChangeProhibitedMessage), convertibleToActionHandler);
         connect(typeid(VolumeChanged), convertibleToActionHandler);
-        connect(typeid(CellularCallAbortedNotification), convertibleToActionHandler);
-        connect(typeid(CellularRingingMessage), convertibleToActionHandler);
-        connect(typeid(CellularCallActiveNotification), convertibleToActionHandler);
-        connect(typeid(CellularHangupCallMessage), convertibleToActionHandler);
     }
 
-    sys::ReturnCodes ApplicationManager::SwitchPowerModeHandler(const sys::ServicePowerMode mode)
+    sys::ReturnCodes ApplicationManagerCommon::SwitchPowerModeHandler(const sys::ServicePowerMode mode)
     {
         LOG_INFO("Power mode: %s", c_str(mode));
 
@@ -567,22 +319,19 @@ namespace app::manager
         return sys::ReturnCodes::Success;
     }
 
-    auto ApplicationManager::startApplication(ApplicationHandle &app) -> bool
+    auto ApplicationManagerCommon::startApplication(ApplicationHandle &app) -> bool
     {
         if (app.state() == ApplicationHandle::State::ACTIVE_BACKGROUND) {
             LOG_INFO("Switching focus to application [%s] (window [%s])", app.name().c_str(), app.switchWindow.c_str());
             setState(State::AwaitingFocusConfirmation);
             app::Application::messageSwitchApplication(
                 this, app.name(), app.switchWindow, std::move(app.switchData), StartupReason::Launch);
+            return true;
         }
-        else {
-            LOG_INFO("Starting application %s", app.name().c_str());
-            app.run(phoneModeObserver->getCurrentPhoneMode(), this);
-        }
-        return true;
+        return false;
     }
 
-    auto ApplicationManager::closeApplications() -> bool
+    auto ApplicationManagerCommon::closeApplications() -> bool
     {
         for (const auto &app : getApplications()) {
             if (app->started()) {
@@ -594,7 +343,7 @@ namespace app::manager
         return true;
     }
 
-    void ApplicationManager::closeNoLongerNeededApplications()
+    void ApplicationManagerCommon::closeNoLongerNeededApplications()
     {
         for (const auto &app : getApplications()) {
             if (app->started() && app->closeable() && !stack.contains(app->name())) {
@@ -604,7 +353,7 @@ namespace app::manager
         }
     }
 
-    auto ApplicationManager::closeApplicationsOnUpdate() -> bool
+    auto ApplicationManagerCommon::closeApplicationsOnUpdate() -> bool
     {
         for (const auto &app : getApplications()) {
             if (app->started()) {
@@ -621,7 +370,7 @@ namespace app::manager
         return true;
     }
 
-    void ApplicationManager::closeApplication(ApplicationHandle *application)
+    void ApplicationManagerCommon::closeApplication(ApplicationHandle *application)
     {
         if (application == nullptr) {
             return;
@@ -636,13 +385,13 @@ namespace app::manager
         application->close();
     }
 
-    auto ApplicationManager::handlePowerSavingModeInit() -> bool
+    auto ApplicationManagerCommon::handlePowerSavingModeInit() -> bool
     {
         LOG_INFO("Going to suspend mode");
         return true;
     }
 
-    auto ApplicationManager::handleSwitchApplication(SwitchRequest *msg, bool closeCurrentlyFocusedApp) -> bool
+    auto ApplicationManagerCommon::handleSwitchApplication(SwitchRequest *msg, bool closeCurrentlyFocusedApp) -> bool
     {
         auto app = getApplication(msg->getName());
         if (app == nullptr) {
@@ -679,9 +428,9 @@ namespace app::manager
         return true;
     }
 
-    void ApplicationManager::onApplicationSwitch(ApplicationHandle &nextApp,
-                                                 std::unique_ptr<gui::SwitchData> &&data,
-                                                 std::string targetWindow)
+    void ApplicationManagerCommon::onApplicationSwitch(ApplicationHandle &nextApp,
+                                                       std::unique_ptr<gui::SwitchData> &&data,
+                                                       std::string targetWindow)
     {
         if (nextApp.name() == rootApplicationName) {
             stack.clear();
@@ -692,7 +441,7 @@ namespace app::manager
         nextApp.switchWindow = std::move(targetWindow);
     }
 
-    void ApplicationManager::requestApplicationClose(ApplicationHandle &app, bool isCloseable)
+    void ApplicationManagerCommon::requestApplicationClose(ApplicationHandle &app, bool isCloseable)
     {
         if (isCloseable) {
             LOG_INFO("Closing application %s", app.name().c_str());
@@ -706,32 +455,13 @@ namespace app::manager
         }
     }
 
-    void ApplicationManager::handleActionRequest(ActionRequest *actionMsg)
+    void ApplicationManagerCommon::handleActionRequest(ActionRequest *actionMsg)
     {
         ActionEntry entry{actionMsg->getAction(), std::move(actionMsg->getData())};
         actionsRegistry.enqueue(std::move(entry));
     }
 
-    void ApplicationManager::handlePhoneModeChanged(sys::phone_modes::PhoneMode phoneMode)
-    {
-        for (const auto app : getStackedApplications()) {
-            changePhoneMode(phoneMode, app);
-        }
-    }
-
-    void ApplicationManager::changePhoneMode(sys::phone_modes::PhoneMode phoneMode, const ApplicationHandle *app)
-    {
-        ActionEntry action{actions::PhoneModeChanged, std::make_unique<gui::PhoneModeParams>(phoneMode)};
-        action.setTargetApplication(app->name());
-        actionsRegistry.enqueue(std::move(action));
-    }
-
-    void ApplicationManager::handleTetheringChanged(sys::phone_modes::Tethering tethering)
-    {
-        notificationProvider.handle(tethering);
-    }
-
-    ActionProcessStatus ApplicationManager::handleAction(ActionEntry &action)
+    ActionProcessStatus ApplicationManagerCommon::handleAction(ActionEntry &action)
     {
         switch (action.actionId) {
         case actions::Home:
@@ -739,7 +469,7 @@ namespace app::manager
         case actions::Launch:
             return handleLaunchAction(action);
         case actions::PhoneModeChanged:
-            return handlePhoneModeChangedAction(action);
+            return ActionProcessStatus::NotHandled;
         case actions::ShowPopup:
             [[fallthrough]];
         case actions::AbortPopup:
@@ -751,42 +481,32 @@ namespace app::manager
         }
     }
 
-    auto ApplicationManager::handleHomeAction(ActionEntry &action) -> ActionProcessStatus
+    auto ApplicationManagerCommon::handleHomeAction(ActionEntry &action) -> ActionProcessStatus
     {
         action.setTargetApplication(resolveHomeApplication());
 
-        SwitchRequest switchRequest(ServiceName, resolveHomeApplication(), resolveHomeWindow(), nullptr);
+        SwitchRequest switchRequest(service::name::appmgr, resolveHomeApplication(), resolveHomeWindow(), nullptr);
         return handleSwitchApplication(&switchRequest) ? ActionProcessStatus::Accepted : ActionProcessStatus::Dropped;
     }
 
-    auto ApplicationManager::resolveHomeWindow() -> std::string
+    auto ApplicationManagerCommon::resolveHomeWindow() -> std::string
     {
-        return phoneLockHandler.isPhoneLocked() ? gui::popup::window::phone_lock_window
-                                                : gui::name::window::main_window;
+        return gui::name::window::main_window;
     }
 
-    auto ApplicationManager::handleOnBoardingFinalize() -> sys::MessagePointer
+    auto ApplicationManagerCommon::handleOnBoardingFinalize() -> sys::MessagePointer
     {
         settings->setValue(settings::SystemProperties::onboardingDone, utils::to_string(true));
         app::manager::Controller::sendAction(this, app::manager::actions::Home);
         return sys::msgHandled();
     }
 
-    auto ApplicationManager::checkOnBoarding() -> bool
+    auto ApplicationManagerCommon::checkOnBoarding() -> bool
     {
         return not utils::getNumericValue<bool>(settings->getValue(settings::SystemProperties::onboardingDone));
     }
 
-    auto ApplicationManager::resolveHomeApplication() -> std::string
-    {
-        if (checkOnBoarding()) {
-            phoneLockHandler.handleUnlockRequest();
-            return app::name_onboarding;
-        }
-        return rootApplicationName;
-    }
-
-    auto ApplicationManager::handleLaunchAction(ActionEntry &action) -> ActionProcessStatus
+    auto ApplicationManagerCommon::handleLaunchAction(ActionEntry &action) -> ActionProcessStatus
     {
         auto launchParams = static_cast<ApplicationLaunchData *>(action.params.get());
         auto targetApp    = getApplication(launchParams->getTargetApplicationName());
@@ -796,27 +516,11 @@ namespace app::manager
 
         targetApp->startupReason = StartupReason::Launch;
         action.setTargetApplication(targetApp->name());
-        SwitchRequest switchRequest(ServiceName, targetApp->name(), gui::name::window::main_window, nullptr);
+        SwitchRequest switchRequest(service::name::appmgr, targetApp->name(), gui::name::window::main_window, nullptr);
         return handleSwitchApplication(&switchRequest) ? ActionProcessStatus::Accepted : ActionProcessStatus::Dropped;
     }
 
-    auto ApplicationManager::handlePhoneModeChangedAction(ActionEntry &action) -> ActionProcessStatus
-    {
-        const auto &targetName = action.target;
-        auto targetApp         = getApplication(targetName);
-        if (targetApp == nullptr || !targetApp->handles(action.actionId)) {
-            return ActionProcessStatus::Dropped;
-        }
-
-        if (targetApp->state() == ApplicationHandle::State::ACTIVE_FORGROUND ||
-            targetApp->state() == ApplicationHandle::State::ACTIVE_BACKGROUND) {
-            app::Application::requestAction(this, targetName, action.actionId, std::move(action.params));
-            return ActionProcessStatus::Accepted;
-        }
-        return ActionProcessStatus::Skipped;
-    }
-
-    auto ApplicationManager::handleActionOnFocusedApp(ActionEntry &action) -> ActionProcessStatus
+    auto ApplicationManagerCommon::handleActionOnFocusedApp(ActionEntry &action) -> ActionProcessStatus
     {
         auto targetApp = getFocusedApplication();
         if (targetApp == nullptr) {
@@ -828,7 +532,7 @@ namespace app::manager
         return ActionProcessStatus::Accepted;
     }
 
-    auto ApplicationManager::handleCustomAction(ActionEntry &action) -> ActionProcessStatus
+    auto ApplicationManagerCommon::handleCustomAction(ActionEntry &action) -> ActionProcessStatus
     {
         const auto actionHandlers = applications.findByAction(action.actionId);
         if (actionHandlers.empty()) {
@@ -860,12 +564,12 @@ namespace app::manager
         const auto focusedAppClose = !(actionPolicy(action.actionId) == Reaction::KeepFocusedAppInBackground ||
                                        (actionParams && actionParams->disableAppClose));
         SwitchRequest switchRequest(
-            ServiceName, targetApp->name(), targetApp->switchWindow, std::move(targetApp->switchData));
+            service::name::appmgr, targetApp->name(), targetApp->switchWindow, std::move(targetApp->switchData));
         return handleSwitchApplication(&switchRequest, focusedAppClose) ? ActionProcessStatus::Accepted
                                                                         : ActionProcessStatus::Skipped;
     }
 
-    auto ApplicationManager::handleCustomActionOnBackgroundApp(ApplicationHandle *app, ActionEntry &action)
+    auto ApplicationManagerCommon::handleCustomActionOnBackgroundApp(ApplicationHandle *app, ActionEntry &action)
         -> ActionProcessStatus
     {
         if (const auto actionFlag = app->actionFlag(action.actionId);
@@ -876,7 +580,7 @@ namespace app::manager
         return ActionProcessStatus::Dropped;
     }
 
-    auto ApplicationManager::handleSwitchBack(SwitchBackRequest *msg) -> bool
+    auto ApplicationManagerCommon::handleSwitchBack(SwitchBackRequest *msg) -> bool
     {
         auto previousApp = getPreviousApplication();
         if (previousApp == nullptr) {
@@ -912,14 +616,14 @@ namespace app::manager
         return true;
     }
 
-    void ApplicationManager::onApplicationSwitchToPrev(ApplicationHandle &previousApp,
-                                                       std::unique_ptr<gui::SwitchData> &&data)
+    void ApplicationManagerCommon::onApplicationSwitchToPrev(ApplicationHandle &previousApp,
+                                                             std::unique_ptr<gui::SwitchData> &&data)
     {
         stack.pop();
         previousApp.switchData = std::move(data);
     }
 
-    auto ApplicationManager::handleInitApplication(ApplicationInitialised *msg) -> bool
+    auto ApplicationManagerCommon::handleInitApplication(ApplicationInitialised *msg) -> bool
     {
         auto app = getApplication(msg->getSenderName());
         if (app == nullptr) {
@@ -939,7 +643,7 @@ namespace app::manager
         return true;
     }
 
-    void ApplicationManager::onApplicationInitialised(ApplicationHandle &app, StartInBackground startInBackground)
+    void ApplicationManagerCommon::onApplicationInitialised(ApplicationHandle &app, StartInBackground startInBackground)
     {
         LOG_DEBUG("Application %s initialised successfully.", app.name().c_str());
 
@@ -963,13 +667,13 @@ namespace app::manager
         }
     }
 
-    void ApplicationManager::onApplicationInitFailure(ApplicationHandle &app)
+    void ApplicationManagerCommon::onApplicationInitFailure(ApplicationHandle &app)
     {
         LOG_ERROR("Failed to initialise %s: Application internal error.", app.name().c_str());
         Controller::switchBack(this);
     }
 
-    auto ApplicationManager::handleDisplayLanguageChange(app::manager::DisplayLanguageChangeRequest *msg) -> bool
+    auto ApplicationManagerCommon::handleDisplayLanguageChange(app::manager::DisplayLanguageChangeRequest *msg) -> bool
     {
         const auto &requestedLanguage = msg->getLanguage();
 
@@ -983,7 +687,7 @@ namespace app::manager
         return true;
     }
 
-    auto ApplicationManager::handleInputLanguageChange(app::manager::InputLanguageChangeRequest *msg) -> bool
+    auto ApplicationManagerCommon::handleInputLanguageChange(app::manager::InputLanguageChangeRequest *msg) -> bool
     {
         const auto &requestedLanguage = msg->getLanguage();
 
@@ -996,17 +700,7 @@ namespace app::manager
         return true;
     }
 
-    auto ApplicationManager::handleDBResponse(db::QueryResponse *msg) -> bool
-    {
-        auto result = msg->getResult();
-        if (auto response = dynamic_cast<db::query::notifications::GetAllResult *>(result.get())) {
-            notificationProvider.handle(response);
-            return true;
-        }
-        return false;
-    }
-
-    void ApplicationManager::rebuildActiveApplications()
+    void ApplicationManagerCommon::rebuildActiveApplications()
     {
         for (const auto &app : getApplications()) {
             if (app && app->valid()) {
@@ -1018,7 +712,7 @@ namespace app::manager
         }
     }
 
-    auto ApplicationManager::handleSwitchConfirmation(SwitchConfirmation *msg) -> bool
+    auto ApplicationManagerCommon::handleSwitchConfirmation(SwitchConfirmation *msg) -> bool
     {
         auto senderApp = getApplication(msg->getSenderName());
         if (senderApp == nullptr) {
@@ -1030,7 +724,7 @@ namespace app::manager
         return onSwitchConfirmed(*senderApp);
     }
 
-    auto ApplicationManager::onSwitchConfirmed(ApplicationHandle &app) -> bool
+    auto ApplicationManagerCommon::onSwitchConfirmed(ApplicationHandle &app) -> bool
     {
         if (getState() == State::AwaitingFocusConfirmation || getState() == State::Running) {
             app.blockClosing = false;
@@ -1054,7 +748,7 @@ namespace app::manager
         return false;
     }
 
-    void ApplicationManager::onLaunchFinished(ApplicationHandle &app)
+    void ApplicationManagerCommon::onLaunchFinished(ApplicationHandle &app)
     {
         // reset startupReason to default Launch
         app.startupReason = StartupReason::Launch;
@@ -1082,7 +776,7 @@ namespace app::manager
         }
     }
 
-    auto ApplicationManager::handleCloseConfirmation(CloseConfirmation *msg) -> bool
+    auto ApplicationManagerCommon::handleCloseConfirmation(CloseConfirmation *msg) -> bool
     {
         auto senderApp = getApplication(msg->getSenderName());
         if (senderApp == nullptr) {
@@ -1093,7 +787,7 @@ namespace app::manager
         return onCloseConfirmed(*senderApp);
     }
 
-    auto ApplicationManager::onCloseConfirmed(ApplicationHandle &app) -> bool
+    auto ApplicationManagerCommon::onCloseConfirmed(ApplicationHandle &app) -> bool
     {
         if (isApplicationCloseable(&app)) {
             app.setState(ApplicationHandle::State::DEACTIVATED);
@@ -1110,7 +804,7 @@ namespace app::manager
         return true;
     }
 
-    auto ApplicationManager::handleMessageAsAction(sys::Message *request) -> std::shared_ptr<sys::ResponseMessage>
+    auto ApplicationManagerCommon::handleMessageAsAction(sys::Message *request) -> std::shared_ptr<sys::ResponseMessage>
     {
         auto actionMsg = dynamic_cast<manager::actions::ConvertibleToAction *>(request);
         if (!actionMsg) {
@@ -1122,73 +816,19 @@ namespace app::manager
         return std::make_shared<sys::ResponseMessage>();
     }
 
-    void ApplicationManager::onPhoneLocked()
-    {
-        if (phoneLockHandler.isPhoneLocked()) {
-            autoLockTimer.stop();
-            return;
-        }
-        if (auto focusedApp = getFocusedApplication(); focusedApp == nullptr || focusedApp->preventsAutoLocking()) {
-            autoLockTimer.start();
-            return;
-        }
-        if (phoneModeObserver->isTetheringOn()) {
-            autoLockTimer.start();
-            return;
-        }
-        if (event::service::api::isTorchOn()) {
-            autoLockTimer.start();
-            return;
-        }
-        phoneLockHandler.handleLockRequest();
-    }
-
-    void ApplicationManager::displayLanguageChanged(std::string value)
+    void ApplicationManagerCommon::displayLanguageChanged(std::string value)
     {
         if (utils::setDisplayLanguage(value)) {
             rebuildActiveApplications();
         }
     }
 
-    void ApplicationManager::lockTimeChanged(std::string value)
-    {
-        if (value.empty()) {
-            LOG_ERROR("No value for auto-locking time period, request ignored");
-            return;
-        }
-        const auto interval = std::chrono::seconds{utils::getNumericValue<unsigned int>(value)};
-        if (interval.count() == 0) {
-            LOG_ERROR("Invalid auto-locking time period of 0s, request ignored");
-            return;
-        }
-        autoLockTimer.restart(interval);
-    }
-
-    auto ApplicationManager::handleAutoLockGetRequest([[maybe_unused]] GetAutoLockTimeoutRequest *request)
-        -> std::shared_ptr<sys::ResponseMessage>
-    {
-        auto intervalValue =
-            settings->getValue(settings::SystemProperties::autoLockTimeInSec, settings::SettingsScope::Global);
-        const auto interval = std::chrono::seconds{utils::getNumericValue<unsigned int>(intervalValue)};
-        return std::make_shared<GetAutoLockTimeoutResponse>(interval);
-    }
-    auto ApplicationManager::handleAutoLockSetRequest(SetAutoLockTimeoutRequest *request)
-        -> std::shared_ptr<sys::ResponseMessage>
-    {
-        auto interval = request->getValue();
-        settings->setValue(settings::SystemProperties::autoLockTimeInSec,
-                           utils::to_string(interval.count()),
-                           settings::SettingsScope::Global);
-        autoLockTimer.restart(interval);
-        return std::make_shared<sys::ResponseMessage>();
-    }
-
-    void ApplicationManager::inputLanguageChanged(std::string value)
+    void ApplicationManagerCommon::inputLanguageChanged(std::string value)
     {
         utils::setInputLanguage(value);
     }
 
-    auto ApplicationManager::handleDOMRequest(sys::Message *request) -> std::shared_ptr<sys::ResponseMessage>
+    auto ApplicationManagerCommon::handleDOMRequest(sys::Message *request) -> std::shared_ptr<sys::ResponseMessage>
     {
         auto app = getFocusedApplication();
         if (app != nullptr) {
@@ -1200,33 +840,8 @@ namespace app::manager
         return std::make_shared<sys::ResponseMessage>(sys::ReturnCodes::Unresolved);
     }
 
-    auto ApplicationManager::handleDeveloperModeRequest(sys::Message *request) -> sys::MessagePointer
+    auto ApplicationManagerCommon::handleDeveloperModeRequest(sys::Message *request) -> sys::MessagePointer
     {
-        if (auto msg = dynamic_cast<sdesktop::developerMode::DeveloperModeRequest *>(request)) {
-            if (dynamic_cast<sdesktop::developerMode::ScreenlockCheckEvent *>(msg->event.get())) {
-                auto response = std::make_shared<sdesktop::developerMode::DeveloperModeRequest>(
-                    std::make_unique<sdesktop::developerMode::ScreenlockCheckEvent>(phoneLockHandler.isPhoneLocked()));
-                bus.sendUnicast(std::move(response), service::name::service_desktop);
-                return sys::msgHandled();
-            }
-        }
         return sys::msgNotHandled();
-    }
-
-    void ApplicationManager::processKeypadBacklightState(bsp::keypad_backlight::State keypadLightState)
-    {
-        auto action = bsp::keypad_backlight::Action::turnOff;
-        switch (keypadLightState) {
-        case bsp::keypad_backlight::State::on:
-            action = bsp::keypad_backlight::Action::turnOn;
-            break;
-        case bsp::keypad_backlight::State::activeMode:
-            action = bsp::keypad_backlight::Action::turnOnActiveMode;
-            break;
-        case bsp::keypad_backlight::State::off:
-            action = bsp::keypad_backlight::Action::turnOff;
-            break;
-        }
-        bus.sendUnicast(std::make_shared<sevm::KeypadBacklightMessage>(action), service::name::evt_manager);
-    }
+    };
 } // namespace app::manager
