+++ products/PurePhone/services/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(evtmgr)
+add_subdirectory(appmgr)
