+++ products/PurePhone/services/appmgr/CMakeLists.txt
@@ -1,0 +1,27 @@
+add_library(appmgr STATIC)
+
+target_sources(appmgr
+    PRIVATE
+        ApplicationManager.cpp
+    PUBLIC
+        include/appmgr/ApplicationManager.hpp
+)
+
+target_include_directories(appmgr
+    PUBLIC
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+)
+
+target_link_libraries(appmgr
+    PRIVATE
+        application-special-input
+        apps-common
+        module-apps
+        module-db
+        module-sys
+        service-appmgr
+        service-cellular
+        service-db
+        service-desktop
+        service-evtmgr
+)
